experiment/pyproject.toml
[project]
name = "color-grid-tool"
version = "0.1.0"
description = "RGB颜色方块拼接工具"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-image>=0.21.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "lpips>=0.1.4",
    "tqdm>=4.65.0",
]
